Refactoring dashboard and css into project table
