+++ nx/blocks/loc/dashboard/dashboard.css
@@ -4,56 +4,6 @@
   margin: 0 auto var(--spacing-500) auto;
 }
 
-.table {
-  display: grid;
-  grid-template-columns: repeat(7, 1fr); /* Adjust number of columns based on your data */
-  border-collapse: collapse;
-  margin-top: 1rem;
-  width: 100%;
-}
-
-.table-header,
-.table-row {
-  display: contents; /* Allows children to behave like grid items */
-}
-
-.table-body,
-.table-row {
-  display: contents;
-}
-
-.table-row {
-  grid-template-columns: repeat(7, 1fr); /* Match number of columns */
-  align-items: stretch; /* Ensures all cells stretch to equal row height */
-  min-height: 50px; /* Set a baseline minimum height */
-}
-
-.table-cell {
-  padding: 8px;
-  border-bottom: 1px solid #e0e0e0;
-  word-wrap: break-word;
-  text-align: left;
-}
-
-.table-header .table-cell {
-  font-weight: bold;
-  background-color: #f3f3f3;
-  text-align: left;
-}
-
-.table-body .table-row:nth-child(even) .table-cell {
-  background-color: #f9f9f9;
-}
-
-.table-body .table-row:hover .table-cell {
-  background-color: #f1f1f1;
-}
-
-.actions {
-  display: flex;
-  gap: 8px;
-}
-
 .icon {
   cursor: pointer;
 }
@@ -75,191 +25,6 @@ h1 {
   background-color: #0056b3;
 }
 
-
-
-.tooltip {
-  margin-left: 5px;
-  display: inline-block;
-  position: relative;
-  text-align: center;
-  font-size: 0.8rem;
-  font-weight: bold;
-  cursor: pointer;
-  width: 24px;
-  height: 24px;
-  line-height: 24px;
-  border-radius: 50%;
-  background-color: black;
-  color: white;
-  box-shadow: 0 2px 4px rgb(0 0 0 / 20%);
-}
-
-.tooltip:hover {
-  background-color: #0056b3;
-}
-
-.tooltip:hover::after {
-  content: attr(title);
-  position: absolute;
-  left: 50%;
-  transform: translateX(-50%);
-  top: calc(100% + 5px);
-  background: #333;
-  color: #fff;
-  padding: 5px 10px;
-  border-radius: 4px;
-  white-space: nowrap;
-  z-index: 10;
-  font-size: 0.9rem;
-  box-shadow: 0 2px 4px rgb(0 0 0 / 20%);
-  visibility: visible;
-  opacity: 1;
-}
-
-.tooltip::after {
-  visibility: hidden;
-  opacity: 0;
-  transition: opacity 0.2s ease-in-out;
-}
-
-.filter-bar {
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-  margin-bottom: 1rem;
-}
-
-.search-input {
-  flex: 1;
-  padding: 8px 12px;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-  font-size: 1rem;
-}
-
-.date-range {
-  display: flex;
-  align-items: center;
-  gap: 0.5rem;
-}
-
-.date-picker {
-  padding: 8px 12px;
-  border: 1px solid #ccc;
-  border-radius: 4px;
-  font-weight: bold;
-}
-
-.date-picker:focus {
-  outline: none;
-  border-color: #007bff;
-}
-
-.search-input:focus {
-  outline: none;
-  border-color: #007bff;
-}
-
-.edit-button {
-  width: 80px; /* Fixed width for all buttons */
-  height: 36px; /* Fixed height for all buttons */
-  padding: 0;
-  border: none;
-  background-color: #007bff;
-  color: #fff;
-  border-radius: 4px;
-  font-size: 0.9rem;
-  cursor: pointer;
-  font-weight: bold;
-  text-align: center;
-  display: inline-flex; /* Ensures content inside the button is centered */
-  justify-content: center;
-  align-items: center;
-  transition: background-color 0.3s ease, transform 0.2s ease;
-}
-
-.edit-button:hover {
-  background-color: #0056b3;
-  transform: translateY(-2px); /* Adds a hover lift effect */
-}
-
-.edit-button:active {
-  background-color: #003d82;
-  transform: translateY(0); /* Reverts the lift effect when clicked */
-}
-
-.edit-button:focus {
-  outline: none;
-  box-shadow: 0 0 0 2px #80bdff; /* Adds a focus ring for accessibility */
-}
-
-.filter-button {
-  padding: 8px 12px;
-  border: 1px solid rgb(204 204 204);
-  background-color: rgb(255 255 255);
-  border-radius: 4px;
-  cursor: pointer;
-}
-
-.filter-popup {
-  position: absolute;
-  background-color: white;
-  border: 1px solid #ddd;
-  border-radius: 4px;
-  box-shadow: 0 2px 6px rgb(0 0 0 / 20%);
-  padding: 16px;
-  z-index: 100;
-  width: 300px;
-}
-
-.filter-popup h3 {
-  margin: 0 0 10px;
-  font-size: 1.1rem;
-}
-
-.filter-section {
-  margin-bottom: 16px;
-}
-
-.filter-section h4 {
-  margin: 0 0 10px;
-  font-size: 1rem;
-  font-weight: bold;
-}
-
-.filter-section label {
-  display: flex;
-  align-items: center;
-  margin-bottom: 8px;
-  cursor: pointer;
-}
-
-.filter-section input[type="checkbox"] {
-  margin-right: 8px;
-}
-
-.filter-section .checkbox-grid {
-  display: grid;
-  grid-template-columns: repeat(2, 1fr); /* Two checkboxes per row */
-  gap: 8px; /* Space between rows and columns */
-}
-
-.apply-filter-button {
-  padding: 10px;
-  border: none;
-  background-color: #007bff;
-  color: #fff;
-  border-radius: 4px;
-  cursor: pointer;
-  font-size: 1rem;
-  width: 100%;
-  text-align: center;
-}
-
-.apply-filter-button:hover {
-  background-color: #0056b3;
-}
-
 .dashboard-header {
   display: flex;
   align-items: center;
@@ -268,64 +33,4 @@ h1 {
   font-size: 1.5rem;
   font-weight: bold;
   color: #333;
-}
-
-.toggle-switch {
-  display: flex;
-  align-items: center;
-  gap: 10px; /* Add spacing between the slider and the label text */
-}
-
-.toggle-switch label {
-  position: relative;
-  display: flex; /* Use flex to align items */
-  align-items: center;
-  gap: 8px; /* Add spacing between slider and text */
-}
-
-.toggle-switch input {
-  opacity: 0;
-  width: 0;
-  height: 0;
-}
-
-.toggle-switch .slider {
-  position: relative;
-  display: inline-block;
-  width: 50px;
-  height: 24px;
-  background-color: #ccc;
-  border-radius: 24px;
-  cursor: pointer;
-  transition: background-color 0.3s ease;
-}
-
-.toggle-switch .slider::before {
-  content: "";
-  position: absolute;
-  height: 18px;
-  width: 18px;
-  left: 3px;
-  bottom: 3px;
-  background-color: white;
-  border-radius: 50%;
-  transition: transform 0.3s ease;
-}
-
-.toggle-switch input:checked + .slider {
-  background-color: #007bff;
-}
-
-.toggle-switch input:checked + .slider::before {
-  transform: translateX(26px);
-}
-
-.toggle-label {
-  font-size: 0.9rem;
-  color: #333;
-  font-weight: bold;
-}
-
-.toggle-switch label:hover .slider {
-  background-color: #0056b3;
 }
