+++ nx/blocks/loc/dashboard/pagination.css
@@ -1,42 +1,42 @@
 .pagination {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin-top: 1rem;
-    gap: 1rem;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin-top: 1rem;
+  gap: 1rem;
 }
 
 .pagination-info {
-    font-size: 0.9rem;
-    color: #555;
+  font-size: 0.9rem;
+  color: #555;
 }
 
 .pagination-controls {
-    display: flex;
-    gap: 0.5rem;
+  display: flex;
+  gap: 0.5rem;
 }
 
 .pagination-btn {
-    padding: 8px 12px;
-    border: 1px solid #ccc;
-    background-color: #fff;
-    border-radius: 4px;
-    cursor: pointer;
+  padding: 8px 12px;
+  border: 1px solid #ccc;
+  background-color: #fff;
+  border-radius: 4px;
+  cursor: pointer;
 }
 
 .pagination-btn:hover {
-    background-color: #f3f3f3;
+  background-color: #f3f3f3;
 }
 
 .pagination-btn.active {
-    background-color: #007bff;
-    color: #fff;
-    font-weight: bold;
-    cursor: default;
+  background-color: #007bff;
+  color: #fff;
+  font-weight: bold;
+  cursor: default;
 }
 
 .pagination-btn:disabled {
-    background-color: #e0e0e0;
-    color: #aaa;
-    cursor: not-allowed;
+  background-color: #e0e0e0;
+  color: #aaa;
+  cursor: not-allowed;
 }
